# Speed up MemPostings.addFor()