@bboreham 2024-06-04
My suggestion would be to check once before calling BinarySearch whether the value inserts at the end, because it is comparatively expensive to set up the search mechanism and it should be very rare in normal operation.