@bwplotka 2025-02-11
Hello from the bug scrub!

As @bboreham mentioned in https://github.com/prometheus/prometheus/pull/13632#issuecomment-2147294284 it still seems this case we are optimizing is rare. I wonder if running prombench would help us. Let us know @prymitive if you want to continue this optimization and need help with starting prombench run (our macrobenchmark framework that runs 2x ~20GB Prometheus servers and compares e.g. against main).